requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
autocommand==2.2.2
backports.csv==1.0.7
backports.tarfile==1.1.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
blis==0.7.11
build==1.2.1
cachecontrol==0.14.0
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cheroot==10.0.1
CherryPy==18.9.0
cleo==2.1.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.1.1
crashtest==0.4.1
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
distlib==0.3.8
Django==4.2.11
dulwich==0.21.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
fastjsonschema==2.19.1
feedparser==6.0.11
filelock==3.13.4
fonttools==4.51.0
future==1.0.0
gensim==4.3.2
idna==3.7
importlib-metadata==7.0.1
importlib-resources==6.4.0
inflect==7.2.0
installer==0.7.0
jaraco.classes==3.3.1
jaraco.collections==5.0.1
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.0
Jinja2==3.1.3
joblib==1.4.0
keyring==24.3.1
kiwisolver==1.4.5
langcodes==3.3.0
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib==3.7.5
more-itertools==10.2.0
msgpack==1.0.8
murmurhash==1.0.10
mysqlclient==2.2.4
nltk==3.8.1
numpy==1.24.4
packaging==23.2
pandas==2.0.3
Pattern==3.6
pdfminer.six==20231228
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.6.0
portend==3.2.0
preshed==3.0.9
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.7.0
pydantic-core==2.18.1
pyparsing==3.1.2
pyproject-hooks==1.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
rapidfuzz==3.6.2
regex==2024.4.16
requests==2.31.0
requests-toolbelt==1.0.0
scikit-learn==1.3.2
scipy==1.10.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
tempora==5.5.1
thinc==8.2.3
threadpoolctl==3.4.0
tomli==2.0.1
tomlkit==0.12.4
tqdm==4.66.2
trove-classifiers==2024.3.3
typeguard==4.2.1
typer==0.9.4
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
wasabi==1.1.2
weasel==0.3.4
wordcloud==1.9.3
zc.lockfile==3.0.post1
zipp==3.17.0
